+++ modules/apps/analytics/analytics-settings-web/src/main/java/com/liferay/analytics/settings/web/internal/portlet/action/BaseAnalyticsMVCActionCommand.java
@@ -1,0 +1,88 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.analytics.settings.web.internal.portlet.action;
+
+import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
+import com.liferay.portal.kernel.security.auth.PrincipalException;
+import com.liferay.portal.kernel.security.permission.PermissionChecker;
+import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
+import com.liferay.portal.kernel.servlet.SessionErrors;
+import com.liferay.portal.kernel.settings.CompanyServiceSettingsLocator;
+import com.liferay.portal.kernel.settings.ModifiableSettings;
+import com.liferay.portal.kernel.settings.Settings;
+import com.liferay.portal.kernel.settings.SettingsFactoryUtil;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.WebKeys;
+
+import javax.portlet.ActionRequest;
+import javax.portlet.ActionResponse;
+import javax.portlet.MutableRenderParameters;
+
+/**
+ * @author Marcellus Tavares
+ */
+public abstract class BaseAnalyticsMVCActionCommand
+	extends BaseMVCActionCommand {
+
+	protected void checkPermissions(ThemeDisplay themeDisplay)
+		throws PrincipalException {
+
+		PermissionChecker permissionChecker =
+			PermissionThreadLocal.getPermissionChecker();
+
+		if (!permissionChecker.isCompanyAdmin(themeDisplay.getCompanyId())) {
+			throw new PrincipalException();
+		}
+	}
+
+	@Override
+	protected void doProcessAction(
+			ActionRequest actionRequest, ActionResponse actionResponse)
+		throws Exception {
+
+		try {
+			ThemeDisplay themeDisplay =
+				(ThemeDisplay)actionRequest.getAttribute(WebKeys.THEME_DISPLAY);
+
+			checkPermissions(themeDisplay);
+
+			storeSettings(actionRequest, getModifiableSettings(themeDisplay));
+		}
+		catch (PrincipalException pe) {
+			SessionErrors.add(actionRequest, pe.getClass());
+
+			MutableRenderParameters mutableRenderParameters =
+				actionResponse.getRenderParameters();
+
+			mutableRenderParameters.setValue("mvcPath", "/error.jsp");
+		}
+	}
+
+	protected ModifiableSettings getModifiableSettings(
+			ThemeDisplay themeDisplay)
+		throws Exception {
+
+		Settings settings = SettingsFactoryUtil.getSettings(
+			new CompanyServiceSettingsLocator(
+				themeDisplay.getCompanyId(), "com.liferay.analytics"));
+
+		return settings.getModifiableSettings();
+	}
+
+	protected abstract void storeSettings(
+			ActionRequest actionRequest, ModifiableSettings modifiableSettings)
+		throws Exception;
+
+}
