+++ modules/apps/analytics/analytics-settings-web/src/main/java/com/liferay/analytics/settings/web/internal/portlet/action/EditSyncedSitesMVCActionCommand.java
@@ -15,26 +15,13 @@
 package com.liferay.analytics.settings.web.internal.portlet.action;
 
 import com.liferay.configuration.admin.constants.ConfigurationAdminPortletKeys;
-import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
-import com.liferay.portal.kernel.security.auth.PrincipalException;
-import com.liferay.portal.kernel.security.permission.PermissionChecker;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.servlet.SessionErrors;
-import com.liferay.portal.kernel.settings.CompanyServiceSettingsLocator;
 import com.liferay.portal.kernel.settings.ModifiableSettings;
-import com.liferay.portal.kernel.settings.Settings;
-import com.liferay.portal.kernel.settings.SettingsFactory;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.WebKeys;
 
 import javax.portlet.ActionRequest;
-import javax.portlet.ActionResponse;
-import javax.portlet.MutableRenderParameters;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Marcellus Tavares
@@ -46,44 +33,12 @@ import org.osgi.service.component.annotations.Reference;
 	},
 	service = MVCActionCommand.class
 )
-public class EditSyncedSitesMVCActionCommand extends BaseMVCActionCommand {
-
-	protected void checkPermissions(ThemeDisplay themeDisplay)
-		throws PrincipalException {
-
-		PermissionChecker permissionChecker =
-			PermissionThreadLocal.getPermissionChecker();
-
-		if (!permissionChecker.isCompanyAdmin(themeDisplay.getCompanyId())) {
-			throw new PrincipalException();
-		}
-	}
+public class EditSyncedSitesMVCActionCommand
+	extends BaseAnalyticsMVCActionCommand {
 
 	@Override
-	protected void doProcessAction(
-			ActionRequest actionRequest, ActionResponse actionResponse)
-		throws Exception {
-
-		try {
-			ThemeDisplay themeDisplay =
-				(ThemeDisplay)actionRequest.getAttribute(WebKeys.THEME_DISPLAY);
-
-			checkPermissions(themeDisplay);
-
-			storeSettings(actionRequest, themeDisplay);
-		}
-		catch (PrincipalException pe) {
-			SessionErrors.add(actionRequest, pe.getClass());
-
-			MutableRenderParameters mutableRenderParameters =
-				actionResponse.getRenderParameters();
-
-			mutableRenderParameters.setValue("mvcPath", "/error.jsp");
-		}
-	}
-
 	protected void storeSettings(
-			ActionRequest actionRequest, ThemeDisplay themeDisplay)
+			ActionRequest actionRequest, ModifiableSettings modifiableSettings)
 		throws Exception {
 
 		String siteReportingGrouping = ParamUtil.getString(
@@ -91,21 +46,11 @@ public class EditSyncedSitesMVCActionCommand extends BaseMVCActionCommand {
 		String[] syncedGroupIds = ParamUtil.getStringValues(
 			actionRequest, "rowIds");
 
-		Settings settings = _settingsFactory.getSettings(
-			new CompanyServiceSettingsLocator(
-				themeDisplay.getCompanyId(), "com.liferay.analytics"));
-
-		ModifiableSettings modifiableSettings =
-			settings.getModifiableSettings();
-
 		modifiableSettings.setValue(
 			"siteReportingGrouping", siteReportingGrouping);
 		modifiableSettings.setValues("syncedGroupIds", syncedGroupIds);
 
 		modifiableSettings.store();
 	}
-
-	@Reference
-	private SettingsFactory _settingsFactory;
 
 }
